fix: 3783 - same visor for mlkit, zxing and awesome
Deleted file:
* `awesome_flash.dart`

Impacted files:
* `smooth_barcode_scanner_awesome.dart`: moved there the "awesome" flash, that is not part of the visor anymore
* `smooth_barcode_scanner_mlkit.dart`: now uses the common visor
* `smooth_barcode_scanner_visor.dart`: moved the "awesome" flash to the awesome scanner
* `smooth_barcode_scanner_zxing.dart`: now uses the common visor
