+++ packages/smooth_app/lib/pages/scan/smooth_barcode_scanner_zxing.dart
@@ -2,13 +2,12 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
 import 'package:qr_code_scanner/qr_code_scanner.dart';
 import 'package:smooth_app/generic_lib/design_constants.dart';
-import 'package:smooth_app/helpers/app_helper.dart';
 import 'package:smooth_app/helpers/camera_helper.dart';
 import 'package:smooth_app/helpers/haptic_feedback_helper.dart';
 import 'package:smooth_app/pages/scan/scan_header.dart';
+import 'package:smooth_app/pages/scan/smooth_barcode_scanner_visor.dart';
 import 'package:smooth_app/themes/constant_icons.dart';
 import 'package:smooth_app/widgets/screen_visibility.dart';
 import 'package:visibility_detector/visibility_detector.dart';
@@ -71,81 +70,65 @@ class _SmoothBarcodeScannerZXingState extends State<SmoothBarcodeScannerZXing> {
           _visible = false;
           _controller?.pauseCamera();
         },
-        child: LayoutBuilder(
-          builder: (BuildContext context, BoxConstraints constraints) => Stack(
-            children: <Widget>[
-              QRView(
-                key: _qrKey,
-                onQRViewCreated: _onQRViewCreated,
-                overlay: QrScannerOverlayShape(
-                  borderColor: Colors.white,
-                  borderRadius: 10,
-                  borderLength: 30,
-                  borderWidth: 10,
-                  cutOutWidth: constraints.maxWidth - 2 * MINIMUM_TOUCH_SIZE,
-                  cutOutHeight: constraints.maxHeight,
-                ),
-                formatsAllowed: _barcodeFormats,
+        child: Stack(
+          children: <Widget>[
+            QRView(
+              key: _qrKey,
+              onQRViewCreated: _onQRViewCreated,
+              formatsAllowed: _barcodeFormats,
+            ),
+            const Center(child: SmoothBarcodeScannerVisor()),
+            const Align(
+              alignment: Alignment.topCenter,
+              child: Text(
+                'ZXing',
+                style: TextStyle(color: Colors.red),
               ),
-              const Align(
-                alignment: Alignment.topCenter,
-                child: Text(
-                  'ZXing',
-                  style: TextStyle(color: Colors.red),
-                ),
-              ),
-              const Align(
-                alignment: Alignment.topCenter,
-                child: ScanHeader(),
-              ),
-              Center(
-                child: SvgPicture.asset(
-                  'assets/icons/visor_icon.svg',
-                  package: AppHelper.APP_PACKAGE,
-                ),
-              ),
-              Align(
-                alignment: Alignment.bottomCenter,
-                child: Row(
-                  mainAxisAlignment: _showFlipCameraButton
-                      ? MainAxisAlignment.spaceBetween
-                      : MainAxisAlignment.end,
-                  crossAxisAlignment: CrossAxisAlignment.center,
-                  children: <Widget>[
-                    if (_showFlipCameraButton)
-                      IconButton(
-                        icon: Icon(ConstantIcons.instance.getCameraFlip()),
+            ),
+            const Align(
+              alignment: Alignment.topCenter,
+              child: ScanHeader(),
+            ),
+            Align(
+              alignment: Alignment.bottomCenter,
+              child: Row(
+                mainAxisAlignment: _showFlipCameraButton
+                    ? MainAxisAlignment.spaceBetween
+                    : MainAxisAlignment.end,
+                crossAxisAlignment: CrossAxisAlignment.center,
+                children: <Widget>[
+                  if (_showFlipCameraButton)
+                    IconButton(
+                      icon: Icon(ConstantIcons.instance.getCameraFlip()),
+                      color: Colors.white,
+                      onPressed: () async {
+                        SmoothHapticFeedback.click();
+                        await _controller?.flipCamera();
+                        setState(() {});
+                      },
+                    ),
+                  FutureBuilder<bool?>(
+                    future: _controller?.getFlashStatus(),
+                    builder: (_, final AsyncSnapshot<bool?> snapshot) {
+                      final bool? flashOn = snapshot.data;
+                      if (flashOn == null) {
+                        return EMPTY_WIDGET;
+                      }
+                      return IconButton(
+                        icon: Icon(flashOn ? Icons.flash_on : Icons.flash_off),
                         color: Colors.white,
                         onPressed: () async {
                           SmoothHapticFeedback.click();
-                          await _controller?.flipCamera();
+                          await _controller?.toggleFlash();
                           setState(() {});
                         },
-                      ),
-                    FutureBuilder<bool?>(
-                      future: _controller?.getFlashStatus(),
-                      builder: (_, final AsyncSnapshot<bool?> snapshot) {
-                        final bool? flashOn = snapshot.data;
-                        if (flashOn == null) {
-                          return EMPTY_WIDGET;
-                        }
-                        return IconButton(
-                          icon:
-                              Icon(flashOn ? Icons.flash_on : Icons.flash_off),
-                          color: Colors.white,
-                          onPressed: () async {
-                            SmoothHapticFeedback.click();
-                            await _controller?.toggleFlash();
-                            setState(() {});
-                          },
-                        );
-                      },
-                    ),
-                  ],
-                ),
+                      );
+                    },
+                  ),
+                ],
               ),
-            ],
-          ),
+            ),
+          ],
         ),
       );
